+++ Back/symfoAPE/src/Controller/NewsController.php
@@ -3,7 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\News;
+use App\Entity\Medium;
 use App\Repository\NewsRepository;
+use App\Utils\Slugger;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -22,7 +26,7 @@ class NewsController extends AbstractController
     /**
      * @Route("/news/{id}", name="news_show", methods={"GET"}))
      */
-    public function one(News $news)
+    public function show(News $news)
     {   
         if (empty($news)) {
             return new JsonResponse(['message' => 'News not found'], Response::HTTP_NOT_FOUND);
@@ -33,9 +37,46 @@ class NewsController extends AbstractController
 
     /**
      * @Route("/api/news/create", name="news_create")
+     */
+    public function create(Request $request, EntityManagerInterface $entityManager, Slugger $slugger) {
+        $newNews = json_decode($request->getContent(), true);  
+ 
+        $news = new News();
+        $news->setTitle($newNews->title);
+        $news->setContent($newNews->content);
+        
+        $slug = $slugger->slugify($news->getTitle());
+        $news->setSlug($slug);
+
+        $entityManager->persist($news);
+        $entityManager->flush();
+        
+        $data = [
+            'status' => 201,
+            'message' => 'Le news à été créé'
+        ];
+        return new JsonResponse($data, 201);  
+    }
+
+    /**
      * @Route("/api/news/{id}/edit", name="news_edit")
      */
-    //TODO
+    public function edit(News $news, Request $request, Slugger $slugger,  EntityManagerInterface $entityManager) 
+    {
+        // $this->denyAccessUnlessGranted
+        $news = json_decode($request->getContent());
+        if ($news === null) {
+            return new JsonResponse(['message' =>'Cette news n\'existe pas.'], Response::HTTP_NOT_FOUND);
+        }
+
+        $slug = $slugger->slugify($news->getTitle());
+        $news->setSlug($slug);
+
+        $entityManager->flush();
+        
+        $response = new JsonResponse($news, 200);
+        return $response;
+    }
 
     /**
      * @Route("/api/news/{id}/delete", name="news_delete")
@@ -44,12 +85,11 @@ class NewsController extends AbstractController
 
         $news = $this->getDoctrine()->getRepository(News::class)->find($id);
         
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($news);
-        $response = new Response();
-        $response->send();
-
-        return $this->json($news);
-      }
-
+            if ($news) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($news);
+                $entityManager->flush();
+            }
+      return new Response(null, 204);
+    }
 }
